表示対象とするファイルの設定項目を追加 ・一定サイズ以上のファイルを除外 ・表示対象とするファイル種別を詳細に指定

+++ DigViewer/PathNode.h
@@ -24,7 +24,7 @@
 @interface PathNodeOmmitingCondition : NSObject
 @property BOOL          isOmmitingRawImage;
 @property NSDictionary* suffixes;
-@property int           maxFileSize;  // it means no limit if negative
+@property long          maxFileSize;  // it means no limit if negative
 - (BOOL) isOmmitingImagePath:(NSString*)path;
 @end
